voyage_search/requirements.txt
Flask==3.0.0
pymongo==4.6.1
voyageai==0.3.7
python-dotenv==1.0.0
Pillow==10.2.0
